+++ examples/event-processor-nfts/CHANGELOG.md
@@ -1,5 +1,12 @@
 # event-processor-nfts
 
+## 0.0.37
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer-js-processor@0.6.11
+
 ## 0.0.36
 
 ### Patch Changes