+++ examples/mud/CHANGELOG.md
@@ -1,5 +1,15 @@
 # web-demo
 
+## 0.1.35
+
+### Patch Changes
+
+- Updated dependencies
+  - ethereum-indexer-js-processor@0.6.11
+  - event-processor-bleeps@0.0.37
+  - event-processor-conquest-eth@0.0.37
+  - event-processor-nfts@0.0.37
+
 ## 0.1.34
 
 ### Patch Changes