+++ examples/web-demo/CHANGELOG.md
@@ -1,5 +1,13 @@
 # web-demo
 
+## 0.1.35
+
+### Patch Changes
+
+- event-processor-bleeps@0.0.37
+- event-processor-conquest-eth@0.0.37
+- event-processor-nfts@0.0.37
+
 ## 0.1.34
 
 ### Patch Changes